Added Player Model & Set Player Camera
Added in the Ashigaru player model for the player and re-organized the files so that the MainCamera is anchored on the model's face. Still having troubles animating the body to move with the camera, so I am going to work on that next. As for now, the model T-Poses, and the head bob clips through the model every now and then. I'll fix that later though.